Minor fixes/improvements
* Fix param_scale

This is a regression of e50e297c6f22dfe806b2d8e3c1266db24758426b. It was
not discovered as this error pops only when the first model parameter is
in INT64.

* Ignore *.pb files

Those are outputs from `exp/original_model_run.py` and
`./build/intermittent-cnn -s`

* Migrate to JupyterLab and add nbdime

For the former, I was initially trying to get a variable inspector. It
turns out it's already better without a plugin like a variable
inspector - for example, tab pages.

The latter is useful for examining ipynb changes.