Reduce startup time on MSP432
In _mainCRTStartup, `.bss` is filled with zeros using memset [1], which
takes a long time as buffers take almost all RAM space. As a workaround,
I put large buffers to another section.

I don't hack the ld script and force __bss_end__ to be the same as
__bss_start__, as __bss_end__ appears to be used to derive the stack
pointer as well.

Using -lc_nano is for reducing the size of `.data` section.

[1] https://sourceware.org/git/?p=newlib-cygwin.git;a=blob;f=newlib/libc/sys/arm/crt0.S;h=64d425900249609d86cd09fe638dc060e6af305b;hb=76bd5cab331a873ac422fdcb7ba5fe79abea94f0#l232